# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend API — Plataforma de Trabajo Independiente On-Demand
  version: 1.0.0
  description: >-

    ## Flujo de Autenticación con Verificación en Dos Pasos (2FA/OTP)


    La API implementa un sistema de autenticación con JWT y verificación OTP de dos pasos.


    ### Flujo de Registro

    1. `POST /auth/register` — Crea el usuario y envía un OTP por Email/SMS.

    2. `POST /auth/verify-otp` — Valida el OTP. Devuelve `accessToken` + `refreshToken`.


    ### Flujo de Login

    1. `POST /auth/login` — Valida credenciales. Si son correctas, envía OTP y retorna estado `PENDING_VERIFICATION`.

    2. `POST /auth/verify-otp` — Valida el OTP. Devuelve `accessToken` + `refreshToken`.


    ### Renovación de token

    - `POST /auth/refresh-token` — Usa el `refreshToken` para obtener un nuevo `accessToken` + nuevo `refreshToken`
    (rotación de token por seguridad).


    ### Tokens

    - **accessToken**: JWT firmado, payload `{ user_id, email, current_role, iat, exp }`. Expira en **1 hora**.

    - **refreshToken**: JWT firmado, payload `{ user_id, jti, exp }`. Expira en **7 días**. Almacenado en BD para
    revocación.


    ### Rate Limiting

    Los endpoints de autenticación tienen un límite de **5 intentos por IP cada 15 minutos**.
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor local de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ingrese el accessToken obtenido de /auth/verify-otp
...
paths:
...
  /workers/{id}/certifications:
    post:
      summary: Subir un documento de certificación
      description: >-
        Permite al proveedor subir un documento (PDF/imagen) de validación para su perfil (BACKGROUND_CHECK,
        ID_VERIFICATION o PROFESSIONAL_LICENSE). Máximo 10MB.
      tags:
        - Certificaciones
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del perfil de proveedor
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document
                - document_type
              properties:
                document:
                  type: string
                  format: binary
                  description: Archivo digitalizado (PDF, JPG, PNG) max 10MB
                document_type:
                  type: string
                  enum:
                    - BACKGROUND_CHECK
                    - ID_VERIFICATION
                    - PROFESSIONAL_LICENSE
                  example: ID_VERIFICATION
      responses:
        '201':
          description: Documento subido y registrado correctamente en estado PENDING
        '400':
          description: Error de validación o formato/tamaño de archivo no permitido
        '401':
          description: No autenticado
        '403':
          description: No autorizado
        '404':
          description: Perfil de trabajador no encontrado
        '500':
          description: Error interno del servidor
    get:
      summary: Listar las certificaciones de un proveedor
      description: Devuelve todos los documentos de certificación subidos por el proveedor asociado.
      tags:
        - Certificaciones
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del perfil de proveedor
      responses:
        '200':
          description: Lista de certificaciones devuelta exitosamente
        '401':
          description: No autenticado
        '403':
          description: No autorizado
        '404':
          description: Perfil de trabajador no encontrado
        '500':
          description: Error interno del servidor
  /certifications/{id}:
    get:
      summary: Obtener el detalle de una certificación
      description: >-
        Retorna la información de una certificación específica. Acceso exclusivo al proveedor propietario y al
        administrador.
      tags:
        - Certificaciones
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID de la certificación
      responses:
        '200':
          description: Detalle de certificación devuelto exitosamente
        '401':
          description: No autenticado
        '403':
          description: No autorizado
        '404':
          description: Certificación no encontrada
        '500':
          description: Error interno del servidor
    patch:
      summary: Reenviar un documento de certificación rechazado
      description: >-
        Permite subir un nuevo archivo para una certificación que fue previamente REJECTED, cambiando su estado a
        PENDING.
      tags:
        - Certificaciones
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID de la certificación
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document
              properties:
                document:
                  type: string
                  format: binary
                  description: Nuevo archivo digitalizado
      responses:
        '200':
          description: Documento reenviado y estado restablecido a PENDING
        '400':
          description: Error de validación o la certificación no se encuentra rechazada
        '401':
          description: No autenticado
        '403':
          description: No autorizado
        '404':
          description: Certificación no encontrada
        '500':
          description: Error interno del servidor
  /certifications/{id}/status:
    patch:
      summary: Actualizar estado de verificación de una certificación (Admin)
      description: >-
        Permite aprobar (APPROVED) o rechazar (REJECTED) una certificación de proveedor. Requiere motivo de rechazo en
        caso de REJECTED.
      tags:
        - Certificaciones
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID de la certificación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verification_status
              properties:
                verification_status:
                  type: string
                  enum:
                    - PENDING
                    - APPROVED
                    - REJECTED
                  example: APPROVED
                rejected_reason:
                  type: string
                  minLength: 5
                  example: Documento de identidad poco legible o borroso
                  description: Requerido solo si status es REJECTED
      responses:
        '200':
          description: Estado de certificación actualizado correctamente
        '400':
          description: Error de validación (ej. falta motivo en rechazo)
        '401':
          description: No autenticado
        '403':
          description: No autorizado
        '404':
          description: Certificación no encontrada
        '500':
          description: Error interno del servidor
...